pyproject.toml
[tool.poetry]
name = "template"
version = "0.0.1"
description = ""
authors = ["Josh Levin <morpheus636@morpheus636.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
python-dotenv = "^0.21.0"
feedparser = "^6.0.10"
yagmail = "^0.15.293"

[tool.poetry.dev-dependencies]
black = "^22.2.0"
flake8 = "^3.9.2"
isort = "^5.9.3"
pytest = "^6.2.4"
pre-commit = "^2.13.0"

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100
include_trailing_comma = "True"
force_grid_wrap = 0
use_parentheses = "True"
combine_as_imports = "True"
force_alphabetical_sort_within_sections = "True"
ensure_newline_before_comments = "True"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
lines_after_imports = 2

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"